Move WindowFlags into their own header
