+++ src/openrct2-ui/interface/Dropdown.h
@@ -14,6 +14,7 @@
 #include <openrct2-ui/UiStringIds.h>
 #include <openrct2/Limits.h>
 #include <openrct2/core/EnumUtils.hpp>
+#include <openrct2/core/FlagHolder.hpp>
 #include <openrct2/core/StringTypes.h>
 #include <openrct2/drawing/Colour.h>
 #include <openrct2/drawing/ImageId.hpp>
